move osc.transport.udp.* up a level to osc.transport
There is a bunch of transport-agnostic logic in osc.transport.udp that should be
shared with the upcoming TCP implementation.

The plan after this is to move UDP-specific things back down into
osc.transport.udp. Then, implement tho corresponding TCP-specific bits in
transport.tcp.
