+++ modules/core/src/main/java/com/illposed/osc/ConsoleEchoServer.java
@@ -9,7 +9,7 @@
 package com.illposed.osc;
 
 import com.illposed.osc.messageselector.JavaRegexAddressMessageSelector;
-import com.illposed.osc.transport.udp.OSCPortIn;
+import com.illposed.osc.transport.OSCPortIn;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import java.io.IOException;
