+++ modules/core/src/main/java/com/illposed/osc/OSCPacketDispatcher.java
@@ -272,7 +272,7 @@ public class OSCPacketDispatcher implements OSCPacketListener {
 	/**
 	 * The work-horse of {@link #handlePacket(OSCPacketEvent)}.
 	 * @param source the origin of the packet, usually an instance of
-	 *   {@link com.illposed.osc.transport.udp.OSCPortIn}
+	 *   {@link com.illposed.osc.transport.OSCPortIn}
 	 * @param packet to be dispatched
 	 * @param timeStamp the associated time-stamp
 	 */
@@ -325,7 +325,7 @@ public class OSCPacketDispatcher implements OSCPacketListener {
 	 * depending on whether it has an associated {@link OSCBundle#getTimestamp() time-stamp},
 	 * and whether we are {@link #isAlwaysDispatchingImmediately() always dispatching immediately}.
 	 * @param source the origin of the packet, usually an instance of
-	 *   {@link com.illposed.osc.transport.udp.OSCPortIn}
+	 *   {@link com.illposed.osc.transport.OSCPortIn}
 	 * @param bundle the bundle to be dispatched
 	 */
 	private void dispatchBundle(final Object source, final OSCBundle bundle) {
@@ -351,7 +351,7 @@ public class OSCPacketDispatcher implements OSCPacketListener {
 	 * depending on whether they have an associated {@link OSCBundle#getTimestamp() time-stamp},
 	 * and whether we are {@link #isAlwaysDispatchingImmediately() always dispatching immediately}.
 	 * @param source the origin of the packet, usually an instance of
-	 *   {@link com.illposed.osc.transport.udp.OSCPortIn}
+	 *   {@link com.illposed.osc.transport.OSCPortIn}
 	 * @param bundle the bundle to be dispatched immediately
 	 */
 	private void dispatchBundleNow(final Object source, final OSCBundle bundle) {
