+++ modules/core/src/main/java/com/illposed/osc/transport/OSCPort.java
@@ -6,7 +6,7 @@
  * See file LICENSE (or LICENSE.html) for more information.
  */
 
-package com.illposed.osc.transport.udp;
+package com.illposed.osc.transport;
 
 import java.io.IOException;
 import java.net.Inet4Address;
