+++ modules/core/src/main/java/com/illposed/osc/transport/package-info.java
@@ -7,6 +7,6 @@
  */
 
 /**
- * Provides means to send and receive OSC content over the UDP(/IP) protocol.
+ * Provides means to send and receive OSC content over a network.
  */
-package com.illposed.osc.transport.udp;
+package com.illposed.osc.transport;
