+++ modules/ui/src/main/java/com/illposed/osc/ui/OscUI.java
@@ -12,8 +12,8 @@ package com.illposed.osc.ui;
 import com.illposed.osc.OSCBundle;
 import com.illposed.osc.OSCMessage;
 import com.illposed.osc.OSCPacket;
-import com.illposed.osc.transport.udp.OSCPort;
-import com.illposed.osc.transport.udp.OSCPortOut;
+import com.illposed.osc.transport.OSCPort;
+import com.illposed.osc.transport.OSCPortOut;
 import com.illposed.osc.argument.OSCTimeTag64;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
